fix: fix visual summary bugs (#131)
Fix a bug where if only cluster B was selected, it was still showing a caption for cluster A below where the representative image would normally be. Also removed all the menus from all Vega visualizations and made sure that some of them display the right scale (in integer steps).