Cleanup for projection fix (#133)
* Remove unused transform in Accumulator

* Make variable local in function since not used outside

* Add comment and try to understand variables

* WIP: Try to cleanup `Accumulator` component initialization

* Cleanup the ZoomSVG component

Since we are not using the transform and zoom level from the event, we don't need to pass them through the event.

* Make X, Y, width, height of Minimap reactive

* WIP: Rewrite Accumulator

Rewrite the Accumulator. Remove unused/ unneeded code, cleanup intialization, if possible makes things reactive. Update culling methods to that the culling is only done after building the hexagons.

* Deactivate the updateQuantizationDataExportFunction in AppView for now

* style: remove local redeclarations of store variables and minor cleanup (#132)

* style: remove local redeclarations of store variables and minor cleanup

* fix: fix initial container size calculation for culling error

Previously, the size of the container on first render was not correctly used to cull hexagons. This has now been fixed.

* fix: fix leftmost column of hexagons positioning

The leftmost column and the second column were all positioned to the far left as of if (!v). This has been fixed.

---------

Co-authored-by: Alex Bäuerle <alex@a13x.io>